changes delete user

+++ src/domain/odm/odm-consumer.js
@@ -73,6 +73,14 @@ export const Create = async ({
 
 export const Delete = async (id) => {
   try {
+    const consumer = await conn.connMongo.Consumer.findById(id)
+    if (!consumer) {
+      throw new Error('Consumer not found')
+    }
+    await conn.connMongo.User.deleteOne({ email: consumer.email })
+    await conn.connMongo.ConsumerGroup.findByIdAndUpdate(consumer.consumerGroup, {
+      $pull: { consumers: consumer._id }
+    })
     return await conn.connMongo.Consumer.deleteOne({ _id: id })
   } catch (error) {
     LogDanger('Cannot Delete consumer', error)
